+++ birder/net/ssl/base.py
@@ -1,3 +1,4 @@
+import copy
 from typing import Any
 from typing import Optional
 from typing import TypedDict
@@ -52,7 +53,11 @@ class SSLBaseNet(nn.Module):
 
         if hasattr(self, "config") is False:  # Avoid overriding registered configs
             self.config = config
-        elif config is not None:
+        else:
+            if self.config is not None:
+                self.config = copy.deepcopy(self.config)  # Avoid mutating registered configs
+
+        if config is not None:
             assert self.config is not None
             self.config.update(config)  # Override with custom config